Move simple challenge management related functions from controllers.challenge to models (fix #43)
